Moved `getCurrentFen` to it's own file
